Update webpack with new defaults, improved code splitting
Updated the `vendor` chunk to no longer include literally everything from node_modules. This results in a chunk that is too big. Instead it will now generate a chunk for the essentials of our app that is required on every page: react, redux, styled-components. Everything else is dynamically sorted by Webpack instead.